Move multiplatform umbrella to new project structure
