+++ multiplatform/umbrella/src/androidMain/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest package="com.gchristov.newsfeed.multiplatform.umbrella" />
